+++ classes/privacy/provider.php
@@ -37,6 +37,7 @@ class provider implements metadata_provider, null_provider {
     public static function get_metadata(collection $collection): collection {
         $collection->add_external_location_link('dteachwhiteboard', [
             'siteurl' => 'privacy:metadata:dteachwhiteboard:siteurl',
+            'licencekey' => 'privacy:metadata:dteachwhiteboard:licencekey',
         ], 'privacy:metadata:dteachwhiteboard');
 
         return $collection;